+++ pkg/catalog/types.go
@@ -17,8 +17,9 @@ type (
 	Schema        = internalcatalog.Schema
 	ModuleEntry   = internalcatalog.ModuleEntry
 	ModuleType    = internalcatalog.ModuleType
-	Variable      = internalcatalog.Variable
-	Output        = internalcatalog.Output
+	Variable          = internalcatalog.Variable
+	VariableReference = internalcatalog.VariableReference
+	Output            = internalcatalog.Output
 	SearchOptions = internalcatalog.SearchOptions
 	SearchResult  = internalcatalog.SearchResult
 	BuildOptions  = internalcatalog.BuildOptions